Updates now that tsl-textures has type: module
