+++ attw.json
@@ -303,7 +303,6 @@
         "tern",
         "throttle-debounce",
         "tryghost__content-api",
-        "tsl-textures",
         "typography",
         "ungap__structured-clone",
         "unpaginated",
